// integrations/netsuite-tba/schema.ts
export const netsuitePaymentUpdateInputSchema = z.object({
    id: z.string(),
    customerId: z.string().optional(),
    amount: z.number().min(0).optional(),
    currency: z.string().optional(),
    description: z.string().optional(),
    paymentReference: z.string().optional(),
    status: z.string().optional(),
    applyTo: z.array(z.string()).optional()
});